+++ src/core/constants.ts
@@ -53,3 +53,7 @@ export const FLOATING_PANEL_OFFSET_X = 80;
 export const FLOATING_PANEL_OFFSET_Y = 80;
 
 export const INFO_PANEL_MIN_WIDTH = 80;
+
+// TODO: (Feature) add as config option
+export const SHOW_UNPOPULATED_REGIONS = false; // Whether to show regions with no demand data by default in the overview panel and layer.
+export const ENFORCE_ONE_DATASET_VISIBLE = true; // Whether to enforce that only one dataset's map layers can be visible at a time, or allow multiple datasets' layers to be toggled on simultaneously